state resource under project instead of commands and setup state permissions
