+++ opal-gwt-client/src/main/java/org/obiba/opal/web/gwt/app/client/magma/presenter/TablePresenter.java
@@ -14,7 +14,6 @@ import com.google.common.collect.Maps;
 import com.google.gwt.core.client.JsArray;
 import com.google.gwt.core.client.JsArrayString;
 import com.google.gwt.core.client.JsonUtils;
-import com.google.gwt.core.shared.GWT;
 import com.google.gwt.http.client.Request;
 import com.google.gwt.http.client.Response;
 import com.google.gwt.user.cellview.client.ColumnSortEvent;
@@ -421,7 +420,7 @@ public class TablePresenter extends PresenterWidget<TablePresenter.Display>
             getView().setFromTables(null, null);
             getView().setWhereScript(null);
           }
-        }, SC_FORBIDDEN, SC_INTERNAL_SERVER_ERROR, SC_NOT_FOUND)//
+        }, SC_FORBIDDEN, SC_INTERNAL_SERVER_ERROR, SC_NOT_FOUND)
         .withCallback(new ViewResourceCallback()).send();
   }
 
@@ -445,7 +444,7 @@ public class TablePresenter extends PresenterWidget<TablePresenter.Display>
 
   private void initProjectCommandsState() {
     ResourceRequestBuilderFactory.newBuilder()
-        .forResource(UriBuilders.PROJECT_COMMANDS_STATE.create().build(table.getDatasourceName()))
+        .forResource(UriBuilders.PROJECT_STATE.create().build(table.getDatasourceName()))
         .withCallback(SC_OK, new ResponseCodeCallback() {
           @Override
           public void onResponseCode(Request request, Response response) {
