+++ opal-shell/src/main/java/org/obiba/opal/web/shell/ProjectCommandsResource.java
@@ -11,7 +11,6 @@ package org.obiba.opal.web.shell;
 
 import com.google.common.collect.Lists;
 import java.util.Arrays;
-import javax.ws.rs.core.Response.ResponseBuilder;
 import org.apache.shiro.SecurityUtils;
 import org.obiba.magma.Datasource;
 import org.obiba.magma.MagmaEngine;
@@ -234,29 +233,6 @@ public class ProjectCommandsResource extends AbstractCommandsResource {
     return launchCommand(refreshCommand);
   }
 
-  @GET
-  @Path("/state")
-  public Response getState() {
-    List<CommandJob> jobs = commandJobService.getHistory().stream()
-        .filter(job -> job.hasProject() && job.getProject()
-            .equals(name)).collect(Collectors.toList());
-
-    boolean isRefreshing = checkCommandIsBlocked(jobs, name, false);
-    boolean isBusy = checkCommandIsBlocked(jobs, name, true);
-
-    ResponseBuilder responseBuilder = Response.ok();
-
-    if (isBusy) {
-      responseBuilder.entity(State.BUSY.name());
-    } else if (isRefreshing) {
-      responseBuilder.entity(State.REFRESHING.name());
-    } else {
-      responseBuilder.entity(State.READY.name());
-    }
-
-    return responseBuilder.build();
-  }
-
   @Override
   protected CommandJob newCommandJob(String jobName, Command<?> command) {
     CommandJob job = super.newCommandJob(jobName, command);
@@ -341,11 +317,5 @@ public class ProjectCommandsResource extends AbstractCommandsResource {
     return Response.created(
         UriBuilder.fromPath("/").path(WebShellResource.class).path(WebShellResource.class, "getCommand").build(jobId))
         .build();
-  }
-
-  public enum State {
-    BUSY, // project has read, write and refresh commands that are pending or being processed
-    READY,
-    REFRESHING // project only has one refresh command being processed (only one should be present at a time)
   }
 }
